feat: Tratamento de erro com ExcepionHandler

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -1,52 +1,71 @@
 package com.devsuperior.dscommerce.services;
 
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DataBaseException;
+import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
 
 	private ProductRepository productRepository;
-	
+
 	public ProductService(ProductRepository productRepository) {
 		this.productRepository = productRepository;
 	}
-	
+
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Product product = productRepository.findById(id).get();
+		Product product = productRepository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado."));
 		return new ProductDTO(product);
 	}
-	
+
 	@Transactional(readOnly = true)
 	public Page<ProductDTO> findAll(Pageable pageable) {
 		return productRepository.findAll(pageable).map(prod -> new ProductDTO(prod));
 	}
-	
+
 	@Transactional
 	public ProductDTO create(ProductDTO productDTO) {
 		Product product = productDTO.converteToEntity();
 		productRepository.save(product);
 		return new ProductDTO(product);
 	}
-	
+
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO productDTO) {
-		Product productEntity = productRepository.findById(id).get();
-		productDTO.converteToEntityUpdate(productEntity);
-		productRepository.save(productEntity);
+		try {
+			Product productEntity = productRepository.getReferenceById(id);
+			productDTO.converteToEntityUpdate(productEntity);
+			productRepository.save(productEntity);
+
+			return new ProductDTO(productEntity);
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado.");
+		}
 		
-		return new ProductDTO(productEntity);
 	}
-	
-	@Transactional
+
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		productRepository.deleteById(id);
+		if(!productRepository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado.");
+		}
+		try {
+			productRepository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DataBaseException("Falha de integridade referencial.");
+		}
 	}
 }
